+++ LibraryCPP/vector.h
@@ -5,7 +5,7 @@
 
 // Vector (dynamic array)
 // Stores integer values inside
-typedef long long Data;
+typedef int Data;
 
 struct Vector;
